+++ code/header/model/sg_models/turbine_governor_model/IEEEG1SB.h
@@ -116,13 +116,13 @@ class IEEEG1SB : public TURBINE_GOVERNOR_MODEL
         FIRST_ORDER_BLOCK fuel_delayer1;
         FIRST_ORDER_BLOCK fuel_delayer2;
         INTEGRAL_BLOCK drum;
-        float Tservo;
-        float Uo, Uc;
-        float K1, K2;
-        float K3, K4;
-        float K5, K6;
-        float K7, K8;
-        float Kb;
+        double Tservo;
+        double Uo, Uc;
+        double K1, K2;
+        double K3, K4;
+        double K5, K6;
+        double K7, K8;
+        double Kb;
 
 };
